pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "podr"
version = "0.1.0"
description = "A CLI tool for cleaning up Kubernetes pods in specific states"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    { name = "Pod Reaper Team" }
]
dependencies = [
    "typer>=0.9.0",
    "kubernetes>=28.1.0",
    "pyyaml>=6.0.1",
]

[project.scripts]
podr = "podr.main:app"

[tool.hatch.build.targets.wheel]
packages = ["podr"]